test: add specs for signing raw messages (#8)
